pyproject.toml
[tool.poetry]
name = "litestar-user-crud"
version = "0.1.0"
description = "REST API for user management with LiteStar and PostgreSQL"
authors = ["Senior Developer"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
python-dotenv = ">=1.0.0"
litestar = {version = ">=2.0.0", extras = ["standard"]}
litestar-granian = "^0"
litestar-asyncpg = "^0"
advanced-alchemy = "^0.20"
msgspec = "^0.18.6"
asyncpg = "^0.30.0"
pydantic-settings = "^2.9.1"
psycopg2-binary = "^2.9.10"


[tool.poetry.group.dev.dependencies]
ruff = "^0.11.6"
black = "^25.1.0"
isort = "^6.0.1"
mypy = "^1.15.0"
pytest = "^8.3.5"
pytest-asyncio = "^0.26.0"
alembic = "^1.15.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
